Modularize persona for v-next
