+++ .gitignore
@@ -67,3 +67,4 @@ fluentui_core/generatedSrc
 fluentui_others/generatedSrc
 fluentui_drawer/generatedSrc
 fluentui_listitem/generatedSrc
+fluentui_persona/generatedSrc
