+++ fluentui_others/src/main/res/values/colors.xml
@@ -11,13 +11,4 @@
     <color name="fluentui_action_bar_indicator_background">@color/fluentui_communication_tint_20</color>
     <color name="fluentui_action_bar_carousel_background">@color/fluentui_communication_blue</color>
 
-    <color name="fluent_presence_available">#6BB700</color>
-    <color name="fluent_presence_away">#FFAA44</color>
-    <color name="fluent_presence_oof">#B4009E</color>
-    <color name="fluent_presence_blocked">#C50F1F</color>
-    <color name="fluent_presence_busy">#C50F1F</color>
-    <color name="fluent_presence_dnd">#C50F1F</color>
-    <color name="fluent_presence_offline">#8A8886</color>
-    <color name="fluent_presence_unknown">#8A8886</color>
-
 </resources>
