+++ fluentui_persona/src/main/java/com/microsoft/fluentui/vNext/persona/Avatar.kt
