+++ fluentui_persona/src/main/res/drawable/ic_fluent_person_48_regular.xml
