+++ fluentui_persona/src/main/res/drawable/ic_fluent_presence_available_16_regular.xml
