+++ fluentui_persona/src/main/res/drawable/ic_fluent_presence_away_16_filled.xml
