+++ fluentui_persona/src/main/res/drawable/ic_fluent_presence_blocked_16_regular.xml
