+++ fluentui_persona/src/main/res/drawable/ic_fluent_presence_busy_16_regular.xml
