+++ fluentui_persona/src/main/res/drawable/ic_fluent_presence_dnd_16_regular.xml
