+++ fluentui_persona/src/main/res/drawable/ic_fluent_presence_offline_16_regular.xml
