+++ fluentui_persona/src/main/res/drawable/ic_fluent_presence_oof_16_regular.xml
